% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_WIG.R
\name{compute_WIG}
\alias{compute_WIG}
\title{Calculating the weigheted information gain(WIG) of each S-genes}
\usage{
compute_WIG(post_prob, path_affected, num_sgene)
}
\arguments{
\item{post_prob}{a list of the probablitiy that an E-gene is affected by an
S-gene with respect to a specific pathway}

\item{path_affected}{a list of the E-genes that are affected by the S-genes
and are realted to a specific pathway}

\item{num_sgene}{a numeric of the number of S-genes}
}
\value{
a list contains the probability of the uniform distribution before
  the nertwork inference, and the WIG profiles of every S-gene,as well as the
  WIGs of S-genes
}
\description{
Calculating the weigheted information gain(WIG) of each S-genes
}
