% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_Tar_greedy.R
\name{nem_Tar_greedy}
\alias{nem_Tar_greedy}
\title{Inferring the signaling network from (epi)genetic aberrations and effect
reporter gene profiles based on greedy hill climbing methods}
\usage{
nem_Tar_greedy(D, Sgenes, S_pattern, initial = NULL, control, verbose = TRUE)
}
\arguments{
\item{D}{an E-gene observation matrix with the dimension of p(number of
samples) × n(number of effect reporter genes)}

\item{Sgenes}{a character vector of S-genes}

\item{S_pattern}{a binary matrix showing the state of whether signaling
components(S-genes)are perturbed, with the dimension of m × m}

\item{initial}{the initial state of the S-gene graph}

\item{control}{a list contains necessary parameters for implementing the
methods,the details could be referred to \code{\link[nem]{nem}}}

\item{verbose}{a logic to whether present the execution process}
}
\value{
a list contains the network inference results, please also refer to
  \code{\link[nem]{nem}}
}
\description{
Inferring the signaling network from (epi)genetic aberrations and effect
reporter gene profiles based on greedy hill climbing methods
}
\seealso{
\code{\link{nem_Tar_triples}} \code{\link[nem]{nem}}
}
