% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nem_Tar_triples.R
\name{nem_Tar_triples}
\alias{nem_Tar_triples}
\title{Inferring the signaling network from (epi)genetic aberrations and effect
reporter gene profiles based on triple realtion inference methods}
\usage{
nem_Tar_triples(D, Sgenes, S_pattern, control, verbose = TRUE)
}
\arguments{
\item{D}{an E-gene observation matrix with the dimension of p(number of
samples) × n(number of effect reporter genes)}

\item{Sgenes}{a character vector of S-genes}

\item{S_pattern}{a binary matrix showing the state of whether signaling
components(S-genes)are perturbed, with the dimension of m × m(m is the
number of S-genes)}

\item{control}{a list contains necessary parameters for implementing the
methods,as nemTar is developed based on nem, the users of nemTar should
apply the function
nem::set.default.parameters(Sgenes,type="mLL",para=para), para is a vector
of length two:false positive rate and false negative rate.}

\item{verbose}{a logic to whether present the execution process}
}
\value{
a list contains the network inference results using triple relation,
  please also refer to the \code{\link[nem]{nem}}
}
\description{
Inferring the signaling network from (epi)genetic aberrations and effect
reporter gene profiles based on triple realtion inference methods
}
\seealso{
\code{\link{nem_Tar_greedy}} \code{\link[nem]{nem}}
}
