% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_post.R
\name{path_post}
\alias{path_post}
\title{To dissect the influenced E-genes and the probalility of influence under the assumption
of nested effects for a specific pathway}
\usage{
path_post(nemTar_rslt, path_list, Sgenes)
}
\arguments{
\item{nemTar_rslt}{the output of the signaling network inferred by nemTar}

\item{path_list}{a character of the the signature genes of an pathway}

\item{Sgenes}{a character of the S-genes within the signaling network}
}
\description{
To dissect the influenced E-genes and the probalility of influence under the assumption
of nested effects for a specific pathway
}
