% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permu_test.R
\name{permu_test}
\alias{permu_test}
\title{Employing the permuation test}
\usage{
permu_test(path_post, nem_rslt, path_list, sampling_times)
}
\arguments{
\item{path_post}{the influenced E-genes and the probalility of influence
under the assumption of nested effects}

\item{nem_rslt}{the output of the signaling network inferred by nemTar}

\item{path_list}{a character of the the signature genes of an pathway}

\item{sampling_times}{a numeric number indicating the times of sampling}
}
\description{
Employing the permuation test
}
