% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIG_double.R
\name{WIG_double}
\alias{WIG_double}
\title{Calculating the weigheted information gain(WIG) of combinational perturbation
of 2 S-genes}
\usage{
WIG_double(path_post, num_sgene)
}
\arguments{
\item{path_post}{the influenced E-genes and the probalility of influence
under the assumption of nested effects}

\item{num_sgene}{a numeric of the number of S-genes}
}
\description{
Calculating the weigheted information gain(WIG) of combinational perturbation
of 2 S-genes
}
