% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIG_sample.R
\name{WIG_sample}
\alias{WIG_sample}
\title{Sampling of the WIG to measure the statistical significance}
\usage{
WIG_sample(path_post, nem_rslt, path_list, num_sgene, sampling_times)
}
\arguments{
\item{path_post}{the influenced E-genes and the probalility of influence
under the assumption of nested effects}

\item{nem_rslt}{the output of the signaling network inferred by nemTar}

\item{path_list}{a character of the the signature genes of an pathway}

\item{num_sgene}{a numeric of the number of S-genes}

\item{sampling_times}{a numeric number indicating the times of sampling}
}
\description{
Sampling of the WIG to measure the statistical significance
}
